+++ crds/tobs/tpn.py
@@ -14,7 +14,7 @@ from __future__ import division
 from __future__ import absolute_import
 import os.path
 
-from crds import rmap, utils
+from crds.core import rmap, utils
 from crds.certify import TpnInfo
 from crds.tobs import TYPES
